通知ベルのコンポーネントの切り出し
